chore: foramtting, fix the 2d/3d .wgsl versioning issue
